Move assets out of Core
