+++ src/assets/index.ts
