+++ src/core/index.ts
@@ -3,7 +3,6 @@
  * @module Module @proc7ts/context-values/core
  */
 export * from './asset';
-export * from './assets';
 export * from './entry';
 export * from './request';
 export * from './static';
