+++ src/core/static/single.entry.spec.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 import { noop } from '@proc7ts/primitives';
+import { cxBuildAsset, cxConstAsset } from '../../assets';
 import { CxBuilder, CxReferenceError } from '../../build';
-import { cxBuildAsset, cxConstAsset } from '../assets';
 import { CxEntry } from '../entry';
 import { CxValues } from '../values';
 import { cxSingle } from './single.entry';
